Fix implementation on GTK and Winforms.
