+++ gtk/src/toga_gtk/factory.py
@@ -31,7 +31,7 @@ from .widgets.table import Table
 from .widgets.textinput import TextInput
 from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import MainWindow, Window
+from .window import DocumentMainWindow, MainWindow, Window
 
 
 def not_implemented(feature):
@@ -75,6 +75,7 @@ __all__ = [
     "Tree",
     "WebView",
     # Windows
+    "DocumentMainWindow",
     "MainWindow",
     "Window",
 ]
