+++ iOS/src/toga_iOS/window.py
@@ -218,10 +218,6 @@ class Window:
 
 
 class MainWindow(Window):
-    def create_menus(self):
-        # No menu handling at present
-        pass
-
     def create_toolbar(self):
         # No toolbar handling at present
         pass
