+++ textual/src/toga_textual/window.py
@@ -212,8 +212,5 @@ class Window:
 
 
 class MainWindow(Window):
-    def create_menus(self):
-        pass
-
     def create_toolbar(self):
         self.interface.factory.not_implemented("Window.create_toolbar()")
