+++ web/src/toga_web/window.py
@@ -122,8 +122,5 @@ class Window:
 
 
 class MainWindow(Window):
-    def create_menus(self):
-        pass
-
     def create_toolbar(self):
         self.interface.factory.not_implemented("Window.create_toolbar()")
